+++ lib/utils.ts
@@ -22,7 +22,6 @@ export const isUserLoggedIn = async () => {
 export const isAdmin = async () => {
   try {
     const user = await account.get();
-    console.log(user);
     return user.labels.includes("admin");
   } catch (e) {
     return false;